idk at this point

+++ src/packages/pinboard/Pinboard.js
@@ -280,9 +280,11 @@ export default class Pinboard {
             while (!message.entry) {
                 await new Promise(resolve => setTimeout(resolve, 100)); // await entry creation before we can delete it
             }
-            message.entry.delete();
-            message.entry = null;
-            message.pinned = false;
+            if (message.entry) {
+                message.entry.delete();
+                message.entry = null;
+                message.pinned = false;
+            }
         }
     }
